+++ src/hotspot/share/cds/cdsConstants.hpp
@@ -28,14 +28,14 @@
 #include "memory/allStatic.hpp"
 typedef struct {
   const char* _name;
-  int _value;
+  size_t _value;
 } CDSConst;
 
 class CDSConstants : AllStatic {
  public:
   static CDSConst offsets[];
   static CDSConst constants[];
-  static int get_cds_constant(const char* name);
-  static int get_cds_offset(const char* name);
+  static size_t get_cds_constant(const char* name);
+  static size_t get_cds_offset(const char* name);
 };
 #endif // SHARE_CDS_CDSCONSTANTS_HPP
